requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
billiard==4.2.0
bleach==6.2.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.2.1
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.13.3
gunicorn==23.0.0
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==11.2.1
pipenv==2023.12.1
pipreqs==0.5.0
platformdirs==4.2.0
prompt_toolkit==3.0.47
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pywin32==310
pyzmq==26.4.0
redis==5.0.8
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
setuptools==69.2.0
six==1.16.0
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
virtualenv==20.25.1
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
